refactor(notify): extract notify pipeline
